Move interpreter test to subdir
