+++ tests/interpreter/arrays_test.go
@@ -1,22 +1,23 @@
-package tests
+package interpreter
 
 import (
 	"strconv"
 	"symbolic_execution_course/testdata"
+	"symbolic_execution_course/tests"
 	"testing"
 )
 
 func TestCompareElement(t *testing.T) {
 	argVariants := [][]int{{-1, 1}, {3, 1}, {1, 6}, {0, 2}, {1, 2}}
-	ctx := PrepareTest("arrays", "CompareElement")
+	ctx := tests.PrepareTest("arrays", "CompareElement")
 
 	for _, variant := range argVariants {
 		testName := "index: " + strconv.Itoa(variant[0]) + ", value: " + strconv.Itoa(variant[1])
 		t.Run(testName, func(t *testing.T) {
 			args := make(map[string]any)
 
-			args["array"] = ArrayArg{
-				elements: []any{1, 2, 3},
+			args["array"] = tests.ArrayArg{
+				Elements: []any{1, 2, 3},
 			}
 
 			args["index"] = variant[0]
@@ -24,8 +25,8 @@ func TestCompareElement(t *testing.T) {
 
 			expected := testdata.CompareElement([]int{1, 2, 3}, variant[0], variant[1])
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -39,7 +40,7 @@ func TestCompareAges(t *testing.T) {
 		{Name: "name3", Age: 10},
 		{Name: "name4", Age: 15},
 	}
-	ctx := PrepareTest("arrays", "CompareAge")
+	ctx := tests.PrepareTest("arrays", "CompareAge")
 
 	for _, variant := range argVariants {
 		testName := "index: " + strconv.Itoa(variant[0]) + ", value: " + strconv.Itoa(variant[1])
@@ -51,7 +52,7 @@ func TestCompareAges(t *testing.T) {
 
 			expected := testdata.CompareAge(people, variant[0], variant[1])
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
 		})
 	}
 }
