+++ tests/interpreter/dynamic_interpretation_test.go
@@ -1,14 +1,15 @@
-package tests
+package interpreter
 
 import (
 	"fmt"
 	"symbolic_execution_course/testdata"
+	"symbolic_execution_course/tests"
 	"testing"
 )
 
 func TestImpossibleBranch(t *testing.T) {
 	args := []int{-5, -4, 0, 2, 5}
-	ctx := PrepareTest("dynamic_interpretation", "ImpossibleBranch")
+	ctx := tests.PrepareTest("dynamic_interpretation", "ImpossibleBranch")
 	for _, a := range args {
 		t.Run(fmt.Sprintf("%d", a), func(t *testing.T) {
 			args := make(map[string]any)
@@ -16,15 +17,15 @@ func TestImpossibleBranch(t *testing.T) {
 
 			expected := testdata.ImpossibleBranch(a)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, -1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, -1, t)
 		})
 	}
 }
 
 func TestRepeatingConditions(t *testing.T) {
 	args := []int{-5, -4, 0, 2, 5}
-	ctx := PrepareTest("dynamic_interpretation", "RepeatingConditions")
+	ctx := tests.PrepareTest("dynamic_interpretation", "RepeatingConditions")
 
 	for _, a := range args {
 		t.Run(fmt.Sprintf("%d", a), func(t *testing.T) {
@@ -33,15 +34,15 @@ func TestRepeatingConditions(t *testing.T) {
 
 			expected := testdata.RepeatingConditions(a)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, -1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, -1, t)
 		})
 	}
 }
 
 func TestImpossibleCondition(t *testing.T) {
 	args := []int{-5, -4, 0, 2, 5}
-	ctx := PrepareTest("dynamic_interpretation", "ImpossibleCondition")
+	ctx := tests.PrepareTest("dynamic_interpretation", "ImpossibleCondition")
 
 	for _, a := range args {
 		t.Run(fmt.Sprintf("%d", a), func(t *testing.T) {
@@ -50,8 +51,8 @@ func TestImpossibleCondition(t *testing.T) {
 
 			expected := testdata.ImpossibleCondition(a)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, -1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, -1, t)
 		})
 	}
 }
