+++ tests/interpreter/function_call_test.go
@@ -1,15 +1,16 @@
-package tests
+package interpreter
 
 import (
 	"strconv"
 	"symbolic_execution_course/testdata"
+	"symbolic_execution_course/tests"
 	"testing"
 )
 
 func TestTwice(t *testing.T) {
 	argVariants := []int{1, 2, -1, -2, 10}
 
-	ctx := PrepareTest("function_call", "Twice")
+	ctx := tests.PrepareTest("function_call", "Twice")
 	for _, variant := range argVariants {
 		t.Run(strconv.Itoa(variant), func(t *testing.T) {
 			args := make(map[string]any)
@@ -18,8 +19,8 @@ func TestTwice(t *testing.T) {
 
 			expected := testdata.Twice(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -27,7 +28,7 @@ func TestTwice(t *testing.T) {
 func TestTwiceComplex(t *testing.T) {
 	argVariants := []complex128{complex(1, 1), complex(2, 2), complex(3, 3)}
 
-	ctx := PrepareTest("function_call", "TwiceComplex")
+	ctx := tests.PrepareTest("function_call", "TwiceComplex")
 	for i, variant := range argVariants {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			args := make(map[string]any)
@@ -36,8 +37,8 @@ func TestTwiceComplex(t *testing.T) {
 
 			expected := testdata.TwiceComplex(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -45,7 +46,7 @@ func TestTwiceComplex(t *testing.T) {
 func TestTwiceStruct(t *testing.T) {
 	argVariants := []int{1, 2, 5, 10}
 
-	ctx := PrepareTest("function_call", "TwiceStruct")
+	ctx := tests.PrepareTest("function_call", "TwiceStruct")
 	for i, variant := range argVariants {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			args := make(map[string]any)
@@ -54,8 +55,8 @@ func TestTwiceStruct(t *testing.T) {
 
 			expected := testdata.TwiceStruct(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -63,7 +64,7 @@ func TestTwiceStruct(t *testing.T) {
 func TestAddRecursive(t *testing.T) {
 	argVariants := [][]int{{1, 1}, {3, 3}, {10, 10}, {20, 20}}
 
-	ctx := PrepareTest("function_call", "AddRecursive")
+	ctx := tests.PrepareTest("function_call", "AddRecursive")
 	for i, variant := range argVariants {
 		t.Run(strconv.Itoa(i), func(t *testing.T) {
 			args := make(map[string]any)
@@ -73,20 +74,20 @@ func TestAddRecursive(t *testing.T) {
 
 			expected := testdata.AddRecursive(variant[0], variant[1])
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
 
 func TestFib(t *testing.T) {
 	args := make(map[string]any)
-	ctx := PrepareTest("function_call", "Fib")
+	ctx := tests.PrepareTest("function_call", "Fib")
 
 	args["n"] = 5
 
 	expected := testdata.Fib(5)
 
-	SymbolicMachineSatTest(ctx, args, expected, t)
-	SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+	tests.SymbolicMachineSatTest(ctx, args, expected, t)
+	tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 }
