+++ tests/interpreter/memory_test.go
@@ -1,15 +1,16 @@
-package tests
+package interpreter
 
 import (
 	"strconv"
 	"symbolic_execution_course/testdata"
+	"symbolic_execution_course/tests"
 	"testing"
 )
 
 func TestStructAllocateStoreRead(t *testing.T) {
 	aVariants := []int{0, 1, 2, 5, 10}
 
-	ctx := PrepareTest("memory", "StructAllocateStoreRead")
+	ctx := tests.PrepareTest("memory", "StructAllocateStoreRead")
 	for _, variant := range aVariants {
 		t.Run(strconv.Itoa(variant), func(t *testing.T) {
 			args := make(map[string]any)
@@ -17,9 +18,9 @@ func TestStructAllocateStoreRead(t *testing.T) {
 
 			expected := testdata.StructAllocateStoreRead(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
-			SymbolicMachineUnsatTest(ctx, args, -2.0, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, -2.0, t)
 		})
 	}
 }
@@ -27,7 +28,7 @@ func TestStructAllocateStoreRead(t *testing.T) {
 func TestStructOfStructAllocateStoreRead(t *testing.T) {
 	aVariants := []int{0, 1, 2, 5, 10}
 
-	ctx := PrepareTest("memory", "StructOfStructAllocateStoreRead")
+	ctx := tests.PrepareTest("memory", "StructOfStructAllocateStoreRead")
 	for _, variant := range aVariants {
 		t.Run(strconv.Itoa(variant), func(t *testing.T) {
 			args := make(map[string]any)
@@ -35,9 +36,9 @@ func TestStructOfStructAllocateStoreRead(t *testing.T) {
 
 			expected := testdata.StructOfStructAllocateStoreRead(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
-			SymbolicMachineUnsatTest(ctx, args, -2.0, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, -2.0, t)
 		})
 	}
 }
@@ -45,7 +46,7 @@ func TestStructOfStructAllocateStoreRead(t *testing.T) {
 func TestArrayAllocateStoreRead(t *testing.T) {
 	aVariants := []int{0, 1, 2, 5, 10}
 
-	ctx := PrepareTest("memory", "ArrayAllocateStoreRead")
+	ctx := tests.PrepareTest("memory", "ArrayAllocateStoreRead")
 	for _, variant := range aVariants {
 		t.Run(strconv.Itoa(variant), func(t *testing.T) {
 			args := make(map[string]any)
@@ -53,8 +54,8 @@ func TestArrayAllocateStoreRead(t *testing.T) {
 
 			expected := testdata.ArrayAllocateStoreRead(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -62,7 +63,7 @@ func TestArrayAllocateStoreRead(t *testing.T) {
 func TestArrayAllocateStoreReadDynamic(t *testing.T) {
 	argVariants := [][]int{{0, 0}, {1, 2}, {5, 10}, {10, 0}, {10, 10}}
 
-	ctx := PrepareTest("memory", "ArrayAllocateStoreReadDynamic")
+	ctx := tests.PrepareTest("memory", "ArrayAllocateStoreReadDynamic")
 	for _, variant := range argVariants {
 		t.Run(strconv.Itoa(variant[0])+"-"+strconv.Itoa(variant[1]), func(t *testing.T) {
 			args := make(map[string]any)
@@ -71,8 +72,8 @@ func TestArrayAllocateStoreReadDynamic(t *testing.T) {
 
 			expected := testdata.ArrayAllocateStoreReadDynamic(variant[0], variant[1])
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -80,11 +81,11 @@ func TestArrayAllocateStoreReadDynamic(t *testing.T) {
 func TestArrayAllocateStoreReadStore(t *testing.T) {
 	args := make(map[string]any)
 
-	ctx := PrepareTest("memory", "ArrayAllocateStoreReadStore")
+	ctx := tests.PrepareTest("memory", "ArrayAllocateStoreReadStore")
 
 	expected := testdata.ArrayAllocateStoreReadStore()
 
-	SymbolicMachineSatTest(ctx, args, expected, t)
-	SymbolicMachineUnsatTest(ctx, args, expected+1, t)
-	SymbolicMachineUnsatTest(ctx, args, -1, t)
+	tests.SymbolicMachineSatTest(ctx, args, expected, t)
+	tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+	tests.SymbolicMachineUnsatTest(ctx, args, -1, t)
 }
