+++ tests/interpreter/overflow_test.go
@@ -1,15 +1,16 @@
-package tests
+package interpreter
 
 import (
 	"strconv"
 	"symbolic_execution_course/testdata"
+	"symbolic_execution_course/tests"
 	"testing"
 )
 
 func TestShortOverflow(t *testing.T) {
 	aVariants := [][]int{{0, 0}, {127, 1}, {127, 5}, {64, 64}, {127, 127}}
 
-	ctx := PrepareTest("overflow", "ShortOverflow")
+	ctx := tests.PrepareTest("overflow", "ShortOverflow")
 	for _, variant := range aVariants {
 		t.Run(strconv.Itoa(variant[0])+"-"+strconv.Itoa(variant[1]), func(t *testing.T) {
 			args := make(map[string]any)
@@ -18,8 +19,8 @@ func TestShortOverflow(t *testing.T) {
 
 			expected := testdata.ShortOverflow(variant[0], variant[1])
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
@@ -27,7 +28,7 @@ func TestShortOverflow(t *testing.T) {
 func TestOverflowInLoop(t *testing.T) {
 	aVariants := []int{0, 10, 20, 30, 40}
 
-	ctx := PrepareTest("overflow", "OverflowInLoop")
+	ctx := tests.PrepareTest("overflow", "OverflowInLoop")
 	for _, variant := range aVariants {
 		t.Run(strconv.Itoa(variant), func(t *testing.T) {
 			args := make(map[string]any)
@@ -35,8 +36,8 @@ func TestOverflowInLoop(t *testing.T) {
 
 			expected := testdata.OverflowInLoop(variant)
 
-			SymbolicMachineSatTest(ctx, args, expected, t)
-			SymbolicMachineUnsatTest(ctx, args, expected+1, t)
+			tests.SymbolicMachineSatTest(ctx, args, expected, t)
+			tests.SymbolicMachineUnsatTest(ctx, args, expected+1, t)
 		})
 	}
 }
